avoid implicit type conversion

+++ src/ColloLogger.cpp
@@ -123,8 +123,7 @@ void ColloLogger::addLog(const size_t size, const char* msg, const LogLevel& lvl
     message += size;
 
     *message++ = '\n';
-
-    mAppendIndex = static_cast<size_t>(message - mAppendBuffer);
+    mAppendIndex = static_cast<unsigned int>(message - mAppendBuffer);
 }
 
 void ColloLogger::swapBuffers()
